User commit

+++ src/main/java/com/example/demo/repository/StudentRepo.java
@@ -3,9 +3,9 @@ package com.example.demo.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.example.studentCrud.entity.StudentEntity;
+import com.example.demo.entity.StudentEntity;
 
 @Repository
-public interface StudentRepo extends JpaRepository<StudentEntity, Integer> {
+public interface StudentRepo extends JpaRepository<StudentEntity, Long> {
     // no code needed here for basic CRUD
 }
